+++ src/ENDFtk/section/2/151/UnresolvedEnergyDependent/JValue/src/generateList.hpp
@@ -1,5 +1,5 @@
 static std::vector< double >
-generateList( int amux, int amun, int amug, int amuf,
+generateList( double amux, double amun, double amug, double amuf,
               std::vector< double >&& energies,
               std::vector< double >&& d,
               std::vector< double >&& gx,
@@ -7,11 +7,7 @@ generateList( int amux, int amun, int amug, int amuf,
               std::vector< double >&& gg,
               std::vector< double >&& gf ) {
 
-  std::vector< double > list{ 0.0, 0.0,
-                              static_cast< double >( amux ),
-                              static_cast< double >( amun ),
-                              static_cast< double >( amug ),
-                              static_cast< double >( amuf ) };
+  std::vector< double > list{ 0.0, 0.0, amux, amun , amug , amuf  };
 
   unsigned int ne = energies.size();
   if ( ( d.size() != ne ) or ( gx.size() != ne ) or
